services/data/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary